% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchyValue.R
\name{hierarchyPath}
\alias{hierarchyPath}
\title{Extract path values at a given hierarchy level}
\usage{
hierarchyPath(value, level_name)
}
\arguments{
\item{value}{Value returned by a hierarchy input (\code{input$...}).}

\item{level_name}{Name of the hierarchy level column.}
}
\value{
A character vector of unique values at that level among
resolved selections.
}
\description{
Extract path values at a given hierarchy level
}
