% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyHierarchy-package.R
\docType{package}
\name{shinyHierarchy-package}
\alias{shinyHierarchy}
\alias{shinyHierarchy-package}
\title{shinyHierarchy: Hierarchical Input Component for Shiny}
\description{
Hierarchical slicer-style input for Shiny applications with multi-selection, tri-state checkboxes, search, cascading selection, and a structured selection value for analytics workflows.
}
\author{
\strong{Maintainer}: Luis Rodriguez \email{devmazter92@gmail.com}

Authors:
\itemize{
  \item Luis Rodriguez \email{devmazter92@gmail.com}
}

}
\keyword{internal}
