% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHierarchyInput.R
\name{updateHierarchyInput}
\alias{updateHierarchyInput}
\title{Change the value of a hierarchy input on the client}
\usage{
updateHierarchyInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.
Default is \code{getDefaultReactiveDomain()}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{choices}{A nested list representing the hierarchy.}

\item{selected}{The selected node id(s).}
}
\description{
Change the value of a hierarchy input on the client
}
